requirements.txt
ruff
Babel
isort
pre-commit

fastapi==0.111.0
uvicorn==0.23.2
loguru==0.7.0
APScheduler==3.10.1
pydantic==2.8.2
aiofiles=24.1.0
pydantic_settings==2.3.4
PyJWT==2.8.0
aerich==0.7.2
aiosmtplib==3.0.1
Jinja2==3.1.3
logfire[fastapi]==0.46.1

redis==5.1.0b7
tortoise-orm[asyncpg]==0.19.*